requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
docopt==0.6.2
greenlet==2.0.2
idna==3.4
numpy==1.23.5
openpyxl==3.0.0
pandas==1.3.5
pipreqs==0.4.13
playwright==1.37.0
pyarrow==10.0.1
pyee==9.0.4
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
yarg==0.1.9
